+++ database/migrations/2020_07_08_224957_create_trx_penjualan_header_table.php
@@ -21,12 +21,12 @@ class CreateTrxPenjualanHeaderTable extends Migration
             $table->string('no_penjualan', 50)->unique();
             $table->string('pembayaran');
             $table->date('tgl_jatuh_tempo')->nullable();
-            $table->double('subtotal');
-            $table->double('discount');
-            $table->double('ppn');
-            $table->double('grandtotal');
-            $table->double('bayar');
-            $table->double('kembali');
+            $table->decimal('subtotal', 14, 2);
+            $table->decimal('discount', 14, 2);
+            $table->decimal('ppn', 14, 2);
+            $table->decimal('grandtotal', 14, 2);
+            $table->decimal('bayar', 14, 2);
+            $table->decimal('kembali', 14, 2);
             $table->text('keterangan')->nullable();
             $table->tinyInteger('status');
             $table->timestampsTz();
